Remove consul-template support. Do not redownload.
Skip download if file we are trying to download already exists on the
disk. We are anyway not deleting SOURCES folder when recreating build
tree.

Remove explicit consul template support as the script we are trying to
make should be generic enough to easily setup packaging for Go-based
projects. Consul template can be easily built by changing the package
name and download root URL at the top of the file. It will make more
sense to expose them as command line arguments/switches.

+++ build.sh
@@ -7,7 +7,6 @@ pkg='consul'
 pkg_version='latest'
 pkg_release='1'
 isDebug=false
-build_consul_template=1
 current_dir=$(dirname ${0})
 build_root=$(realpath $current_dir/rpmbuild)
 download_url_root='https://releases.hashicorp.com/consul'
@@ -48,7 +47,6 @@ function usage()
     echo -e "  \tbuilding this package for your company, you can set this"
     echo -e "  \tto company name. Default: 1";
     echo -e "\n-b\tPath that will contain RPM build tree. Default is current dir.";
-    echo -e "\n-t\tAlso build consul-template RPM. Does not build by default.";
     echo -e "\n-l\tList available versions.";
     echo -e "\n-h\tShow this help message and exit.";
     echo -e "\n-d\tPrint debugging statements.";
@@ -57,8 +55,8 @@ function usage()
 
 function parse_command()
 {
-    SHORT=v:r:b:lthd
-    LONG=version:,release:,build_root:,list,template,help,debug
+    SHORT=v:r:b:lhd
+    LONG=version:,release:,build_root:,list,help,debug
     PARSED=$(getopt --options $SHORT --longoptions $LONG --name "$0" -- "$@")
     if [[ $? -ne 0 ]]; then
         # e.g. $? == 1
@@ -97,10 +95,6 @@ function parse_command()
                 echo "Available versions:"
                 print_available_versions
                 exit 0
-                shift
-                ;;
-            -t|--template)
-                build_consul_template=0
                 shift
                 ;;
             -h|--help)
@@ -213,6 +207,9 @@ function download_and_verify()
     sources_dir="$build_root/SOURCES"
 
     for file in $core_archive_name $ui_archive_name $checksum; do
+        # Skip download if file already exists
+        [ -f "$sources_dir/$file" ] && continue
+
         dl_url="$download_url_root/$pkg_version/$file"
         print_debug_line "${FUNCNAME[0]} : Downloading $dl_url to $sources_dir/$file"
         wget --max-redirect=0 -O $sources_dir/$file $dl_url &>/dev/null
@@ -251,6 +248,3 @@ download_and_verify
 # _topdir and _tmppath are magic rpm variables that can be defined in ~/.rpmmacros
 # For ease of reliable builds they are defined here on the command line.
 rpmbuild -ba --define="_topdir $build_root" --define="buildroot $build_root/BUILDROOT" --define="pkg_version $pkg_version" --define="rpm_release $pkg_release" $build_root/SPECS/$pkg.spec
-if [ "$build_consul_template" -eq "0" ]; then
-    rpmbuild -ba --define="_topdir $build_root" --define="buildroot $build_root/BUILDROOT" --define="pkg_version $pkg_version" --define="rpm_release $pkg_release" $build_root/SPECS/$pkg-template.spec
-fi
